requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-chroma
langchain-huggingface
langchain-google-genai
langchain-community
langchain-text-splitters
sentence-transformers
chromadb
pydantic